gdbserver: adjust pid after the target attaches
The PID argument of the attach operation may be interpreted in a
target specific way, and may not necessarily mean that the process we
attached to has PID as its process id.  This is for example the case
for Intel GT targets, where the PID refers to the process id of the
host application that is using the devices.  The target in that case
creates a new process for each device where the PID of the process is
the device id/ordinal.  Therefore, once the target completes the
attach, we want adjust the PID value.  For this, revise the meaning
of the return value of the 'attach' target op in case of success:
return the PID of the process that was attached to.

The PID argument to the 'attach' target op is of type 'int' in caller
sites.  Also, the type of the PID field in ptid_t is of type 'int'.
Hence, update the parameter of 'attach' to 'int', too.

+++ gdbserver/win32-low.h
@@ -96,7 +96,7 @@ public:
   int create_inferior (const char *program,
 		       const std::vector<char *> &program_args) override;
 
-  int attach (unsigned long pid) override;
+  int attach (int pid) override;
 
   int kill (process_info *proc) override;
